third_party/blink/web_tests/virtual/css-anchor-update-enabled/external/wpt/css/css-anchor-position/position-area-scrolling-007-expected.txt

This is a testharness.js-based test.
[FAIL] Scroll down to 0,10
  assert_equals: expected 160 but got 1000
[FAIL] Scroll right to 10,10
  assert_equals: expected 10 but got 250
[FAIL] Scroll up to 10,0
  assert_equals: expected 10 but got 250
[FAIL] Scroll down to 10,20
  assert_equals: expected 10 but got 250
[FAIL] Scroll down to 101,20
  assert_equals: expected 20 but got 250
[FAIL] Scroll down to 101,30
  assert_equals: expected 251 but got 750
Harness: the test ran to completion.
